Bug 1366909 - Integrate the test-view code into Treeherder
Modify the code to:
* share assets and global settings wherever possible
* update links going both directions
* other small UI tweaks for uniformity with Treehreder
* Fixed a few routing dead-ends on the react side
* Removed the dead TestDetail file we weren't using anyway
