+++ neutrino-custom/base.js
@@ -161,11 +161,11 @@ module.exports = neutrino => {
             xhtml: true,
             mobile: true,
             minify: HTML_MINIFY_OPTIONS,
-            title: "Treeherder TestGroup - Experimental",
+            title: "Treeherder TestView",
             meta: [
                 {
                     "name": "description",
-                    "content": "Treeherder TestGroup - Experimental"
+                    "content": "Treeherder TestView"
                 },
                 {
                     "name": "author",
@@ -211,6 +211,9 @@ module.exports = neutrino => {
             logViewerApp: require.resolve(path.join(UI, 'js/logviewer.js')),
             userguideApp: require.resolve(path.join(UI, 'js/userguide.js'))
         });
+
+    neutrino.config.devtool('source-map');
+
 };
 
 module.exports.CWD = CWD;
