+++ neutrino-custom/test.js
@@ -7,27 +7,13 @@ module.exports = neutrino => {
     basePreset(neutrino);
     karmaPreset(neutrino);
 
-    // Add an isntanbul loader to generate coverage for js(x) in ui/
-    neutrino.config.module
-        .rule('coverage')
-        .post()
-        .include(UI)
-        .test(/\.jsx?$/)
-        .loader('istanbul', require.resolve('istanbul-instrumenter-loader'));
-
     // Normal karma config
     neutrino.custom.karma = {
         browsers: ['Firefox'],
-        coverageIstanbulReporter: {
-            reports: ['html'],
-            fixWebpackSourcePaths: true
-        },
         plugins: [
             require.resolve('karma-webpack'),
             require.resolve('karma-firefox-launcher'),
-            require.resolve('karma-coverage'),
             require.resolve('karma-jasmine'),
-            require.resolve('karma-coverage-istanbul-reporter')
         ],
         frameworks: ['jasmine'],
         files: [
@@ -37,6 +23,5 @@ module.exports = neutrino => {
         preprocessors: {
             'tests/ui/unit/init.js': ['webpack'],
         },
-        reporters: ['progress', 'coverage-istanbul'],
     };
 };
