+++ package.json
@@ -24,8 +24,10 @@
     "angular1-ui-bootstrap4": "2.4.22",
     "bootstrap": "4.0.0-beta.2",
     "deepmerge": "1.5.2",
+    "eslint-plugin-react": "7.5.1",
     "font-awesome": "4.7.0",
     "hawk": "6.0.2",
+    "history": "4.7.2",
     "jquery": "3.2.1",
     "jquery.scrollto": "2.1.2",
     "js-yaml": "3.10.0",
@@ -44,29 +46,24 @@
     "raw-loader": "0.5.1",
     "react": "15.6.2",
     "react-dom": "15.6.2",
-    "react-fontawesome": "^1.5.0",
-    "react-hot-loader": "3.0.0-beta.6",
-    "react-redux": "5.0.3",
-    "react-router": "4.1.2",
-    "react-router-dom": "4.1.2",
-    "react-transition-group": "1.1.2",
-    "reactstrap": "4.3.0",
-    "redux": "3.6.0",
+    "react-fontawesome": "1.6.1",
+    "react-hot-loader": "3.1.3",
+    "react-redux": "5.0.6",
+    "react-router-dom": "4.2.2",
+    "reactstrap": "5.0.0-alpha.4",
+    "redux": "3.7.2",
     "redux-debounce": "1.0.1",
-    "superagent-promise": "1.1.0",
     "taskcluster-client": "2.5.4"
   },
   "devDependencies": {
     "angular-mocks": "1.5.11",
     "enzyme": "3.1.1",
     "enzyme-adapter-react-15": "1.0.5",
-    "react-test-renderer": "15.6.2",
-    "istanbul-instrumenter-loader": "2.0.0",
     "jasmine-core": "2.8.0",
-    "karma-coverage-istanbul-reporter": "1.3.0",
     "karma-firefox-launcher": "1.0.1",
     "karma-jasmine": "1.1.0",
-    "neutrino-preset-karma": "4.2.1"
+    "neutrino-preset-karma": "4.2.1",
+    "react-test-renderer": "15.6.2"
   },
   "scripts": {
     "build": "node ./node_modules/neutrino/bin/neutrino build --presets ./neutrino-custom/production.js",
