+++ ui/partials/main/jobs.html
@@ -26,7 +26,7 @@
     <th-result-counts class="result-counts"></th-result-counts>
     <span class="result-set-buttons">
       <a class="btn btn-sm btn-resultset test-view-btn"
-         href="{{testBasedUIHost}}?repo={{repoName}}&revision={{resultset.revision}}"
+         href="/testview.html?repo={{repoName}}&revision={{resultset.revision}}"
          target="_blank"
          title="View details on failed test results for this push">View Tests</a>
       <button class="btn btn-sm btn-resultset cancel-all-jobs-btn"
