+++ ui/partials/main/thGlobalTopNavPanel.html
@@ -5,7 +5,8 @@
     <span class="dropdown">
       <button id="th-logo" title="Treeherder services" role="button"
               data-toggle="dropdown"
-              class="btn btn-view-nav dropdown-toggle">Treeherder
+              class="btn btn-view-nav dropdown-toggle">
+        <img src="../../img/treeherder-logo.png" alt="Treeherder" />
       </button>
       <ul class="dropdown-menu" role="menu" aria-labelledby="th-logo">
         <li><a href="perf.html" class="dropdown-item">Perfherder</a></li>
