+++ ui/test-view/redux/configureStore.js
@@ -5,15 +5,9 @@ import {
   applyMiddleware,
 } from 'redux';
 import * as groupsStore from './modules/groups';
-import { platformMap } from './store'
+import { platformMap } from '../../js/constants';
 import createHistory from 'history/createBrowserHistory';
-import createDebounce from 'redux-debounce'
-
-// // TODO: When this app is part of Treeherder, we can get this from the router / URL
-export const treeherder_host = process.env.TREEHERDER_HOST ? process.env.TREEHERDER_HOST : 'treeherder.mozilla.org';
-export const fetch_protocol = process.env.FETCH_PROTOCOL ? process.env.FETCH_PROTOCOL : 'https';
-export const treeherder = `${fetch_protocol}://${treeherder_host}`;
-const request = (url, options) => fetch(url.replace('http:', `${fetch_protocol}:`), options);
+import createDebounce from 'redux-debounce';
 
 function getGroupText(group) {
   const symbol = group.symbol.startsWith('tc-') ?
@@ -26,10 +20,10 @@ function getId(hash) {
   return atob(hash).split(':')[1];
 }
 
-async function fetchTests(store, fetchOptions) {
+async function fetchTests(store, fetchParams) {
   let fetchStatus = 'No failed tests to show';
-  const { url, filter, options, hideClassified, bugSuggestions } = fetchOptions;
-  const response = await request(url, fetchOptions);
+  const { url, filter, options, hideClassified, bugSuggestions } = fetchParams;
+  const response = await fetch(url, fetchParams);
   const pushResp = await response.json();
   // Here the json is the job_detail result.
   // We need to take each entry and query for the errorsummary.log
@@ -41,7 +35,7 @@ async function fetchTests(store, fetchOptions) {
     id: getId(pushData.id),
     repository: pushData.repository,
   };
-  let payload = {groups: {}, rowData: {}, fetchStatus, push, treeherder, filter };
+  let payload = { groups: {}, rowData: {}, fetchStatus, push, filter };
   const jobs = pushData.jobs.edges;
   // It's possible there are no failed jobs, so return early.
 
@@ -140,9 +134,9 @@ function buildFailureLines(lsAcc, failureLine) {
   return lsAcc;
 }
 
-async function fetchOptions(store, fetchOptions) {
-  const { url } = fetchOptions;
-  const resp = await request(url, fetchOptions);
+async function fetchOptions(store, fetchParams) {
+  const { url } = fetchParams;
+  const resp = await fetch(url, fetchParams);
   const options = await resp.json();
   store.dispatch({
     type: groupsStore.types.STORE_OPTIONS,
@@ -155,9 +149,9 @@ async function fetchOptions(store, fetchOptions) {
   });
 }
 
-async function fetchCounts(store, fetchOptions) {
-  const { url } = fetchOptions;
-  const resp = await request(url, fetchOptions);
+async function fetchCounts(store, fetchParams) {
+  const { url } = fetchParams;
+  const resp = await fetch(url, fetchParams);
   const pushStatus = await resp.json();
   const counts = {
     success: 0,
@@ -248,8 +242,8 @@ function filterTests(store, { filter, groups, options, hideClassified }) {
   });
 }
 
-function toggleHideClassified(store, fetchOptions) {
-  const { filter, groups, options, hideClassified } = fetchOptions;
+function toggleHideClassified(store, fetchParams) {
+  const { filter, groups, options, hideClassified } = fetchParams;
   const rowData = filterGroups(filter, groups, options, hideClassified);
 
   store.dispatch({
@@ -271,12 +265,12 @@ async function fetchBugs(store, { rowData, url }) {
   }, {});
   // Do a request for each url in the keys of the testMap.  Using them as keys eliminates
   // duplicate requests since multiple tests can be in the same job.
-  const responses = await Promise.all(Object.keys(testMap).map(url => request(url)));
+  const responses = await Promise.all(Object.keys(testMap).map(url => fetch(url)));
   const respData = await Promise.all(responses.map(promise => promise.json()));
   const bugSuggestions = respData.reduce((bsAcc, data, idx) => {
     testMap[responses[idx].url].forEach((test) => {
       test.bugs = { ...test.bugs, ...extractBugSuggestions(data, test.name) };
-      bsAcc = { ...bugSuggestions, [`${test.jobGroup}-${test.name}`]: test.bugs};
+      bsAcc = { ...bsAcc, [`${test.jobGroup}-${test.name}`]: test.bugs };
     });
     return bsAcc;
   }, {});
