+++ ui/test-view/redux/store.js
@@ -1,88 +1,3 @@
 import configureStore from './configureStore';
 
 export const { store, actions } = configureStore();
-// Map platform key to a user-friendly platform name
-export const platformMap = {
-  'linux32': 'Linux',
-  'linux32-qr': 'Linux QuantumRender',
-  'linux32-nightly': 'Linux Nightly',
-  'linux64': 'Linux x64',
-  'linux64-qr': 'Linux x64 QuantumRender',
-  'linux64-nightly': 'Linux x64 Nightly',
-  'linux64-stylo': 'Linux x64 Stylo',
-  'osx-10-6': 'OS X 10.6',
-  'osx-10-7': 'OS X 10.7',
-  'osx-10-8': 'OS X 10.8',
-  'osx-10-9': 'OS X 10.9',
-  'osx-10-10': 'OS X 10.10',
-  'osx-10-11': 'OS X 10.11',
-  'osx-10-7-qr': 'OS X 10.7 QuantumRender',
-  'windowsxp': 'Windows XP',
-  'windows7-32': 'Windows 7',
-  'windows7-32-vm': 'Windows 7 VM',
-  'windows7-64': 'Windows 7 x64',
-  'windows8-32': 'Windows 8',
-  'windows8-64': 'Windows 8 x64',
-  'windows10-32': 'Windows 10',
-  'windows10-64': 'Windows 10 x64',
-  'windows10-64-vm': 'Windows 10 x64 VM',
-  'windows2012-32': 'Windows 2012',
-  'windows2012-32-qr': 'Windows 2012 QuantumRender',
-  'windows2012-64': 'Windows 2012 x64',
-  'windows2012-64-qr': 'Windows 2012 x64 QuantumRender',
-
-  'android-2-2-armv6': 'Android 2.2 Armv6',
-  'android-2-2': 'Android 2.2',
-  'android-2-3-armv6': 'Android 2.3 Armv6',
-  'android-2-3': 'Android 2.3',
-  'android-2-3-armv7-api9': 'Android 2.3 API9',
-  'android-4-0': 'Android 4.0',
-  'android-4-0-armv7-api10': 'Android 4.0 API10+',
-  'android-4-0-armv7-api11': 'Android 4.0 API11+',
-  'android-4-0-armv7-api15': 'Android 4.0 API15+',
-  'android-4-2-x86': 'Android 4.2 x86',
-  'android-4-2': 'Android 4.2',
-  'android-4-2-armv7-api11': 'Android 4.2 API11+',
-  'android-4-2-armv7-api15': 'Android 4.2 API15+',
-  'android-4-3': 'Android 4.3',
-  'android-4-3-armv7-api11': 'Android 4.3 API11+',
-  'android-4-3-armv7-api15': 'Android 4.3 API15+',
-  'android-4-4': 'Android 4.4',
-  'android-4-4-armv7-api11': 'Android 4.4 API11+',
-  'android-4-4-armv7-api15': 'Android 4.4 API15+',
-  'android-5-0-armv7-api11': 'Android 5.0 API11+',
-  'android-5-0-armv7-api15': 'Android 5.0 API15+',
-  'android-5-1-armv7-api15': 'Android 5.1 API15+',
-  'android-6-0-armv8-api15': 'Android 6.0 API15+',
-  'android-7-1-armv8-api15': 'Android 7.1 API15+',
-  'b2gdroid-4-0-armv7-api11': 'B2GDroid 4.0 API11+',
-  'b2gdroid-4-0-armv7-api15': 'B2GDroid 4.0 API15+',
-  'android-4-0-armv7-api11-partner1': 'Android API11+ partner1',
-  'android-4-0-armv7-api15-partner1': 'Android API15+ partner1',
-  'android-api-15-gradle': 'Android API15+ Gradle',
-
-  'b2g-linux32': 'B2G Desktop Linux',
-  'b2g-linux64': 'B2G Desktop Linux x64',
-  'b2g-osx': 'B2G Desktop OS X',
-  'b2g-win32': 'B2G Desktop Windows',
-  'b2g-emu-ics': 'B2G ICS Emulator',
-  'b2g-emu-jb': 'B2G JB Emulator',
-  'b2g-emu-kk': 'B2G KK Emulator',
-  'b2g-emu-x86-kk': 'B2G KK Emulator x86',
-  'b2g-emu-l': 'B2G L Emulator',
-  'b2g-device-image' : 'B2G Device Image',
-  'mulet-linux32' : 'Mulet Linux',
-  'mulet-linux64' : 'Mulet Linux x64',
-  'mulet-osx': 'Mulet OS X',
-  'mulet-win32': 'Mulet Windows',
-
-  'graphene-linux64': 'Graphene Linux x64',
-  'graphene-osx': 'Graphene OS X',
-  'graphene-win64': 'Graphene Windows x64',
-  'horizon-linux64': 'Horizon Linux x64',
-  'horizon-osx': 'Horizon OS X',
-  'horizon-win64': 'Horizon Windows x64',
-
-  'gecko-decision': 'Gecko Decision Task',
-  'other': 'Other'
-};
