+++ ui/test-view/ui/App.js
@@ -2,7 +2,6 @@ import React from 'react';
 import { BrowserRouter, Route, Switch } from 'react-router-dom';
 import Navigation from './Navigation';
 import Groups from './Groups';
-import Test from './TestDetail';
 import NotFound from './NotFound';
 
 function hasProps(search) {
@@ -17,11 +16,10 @@ const App = () => (
       <Navigation />
       <main>
         <Switch>
-          <Route exact path="/" render={props => (hasProps(props.location.search) ?
+          <Route exact path="/testview.html" render={props => (hasProps(props.location.search) ?
             <Groups {...props} /> :
             <NotFound {...props} />)} />
           <Route name="search" path="?revision=:revision&repo=:repo" handler={Groups} />
-          <Route exact path="/:pushId/tests/:testId" component={Test} />
           <Route name="notfound" component={NotFound} />
         </Switch>
       </main>
