+++ ui/test-view/ui/Navigation.js
@@ -1,6 +1,6 @@
 import React from 'react';
-import { Badge, Nav, Navbar, NavbarToggler, NavbarBrand, Collapse } from 'reactstrap';
-import logoUrl from '../img/logo.png';
+import { Nav, Navbar, Collapse } from 'reactstrap';
+import logoUrl from '../../img/treeherder-logo.png';
 
 export default class Navigation extends React.Component {
   constructor(props) {
@@ -14,14 +14,10 @@ export default class Navigation extends React.Component {
 
   render() {
     return (
-      <Navbar toggleable fixed="top" light style={{ borderBottom: '1px solid #ddd', backgroundColor: '#fff' }}>
-        <NavbarToggler right onClick={() => this.toggle()} />
-        <NavbarBrand href="/">
-          <img src={logoUrl} className="img-fluid" style={{ maxHeight: 26, verticalAlign: 'baseline' }} />
-          <Badge size="sm" color="info" style={{ marginLeft: '1rem', verticalAlign: 'top', marginTop: '.25rem' }}>
-            experimental
-          </Badge>
-        </NavbarBrand>
+      <Navbar expand fixed="top" className="th-top-navbar">
+        <a id="th-logo" href="/">
+          <img src={logoUrl} alt="Treeherder" />
+        </a>
         <Collapse isOpen={this.state.isOpen} navbar>
           <Nav navbar className="mr-auto" />
         </Collapse>
